Split legacy styled-components exports into separate files
Addresses Copilot review comments by separating plain CSS components from
legacy styled-components exports to improve tree-shaking and reduce dependencies.

Changes:
- Created Headings.legacy.ts with legacy styled-components exports (h3Style,
  h4Style, h4DesktopStyle, h4MobileStyle, h3MobileFontSize, h3MobileLineHeight)
- Created Typography.legacy.ts with legacy styled-components exports (textStyle,
  textRegularStyle, bodyCopyRegularStyle, linkStyle, decoratedLinkStyle, etc.)
- Removed styled-components imports from Headings.tsx (now styled-components-free)
- Updated index.ts to export from legacy files instead of defining exports inline
- Removed unused CSS imports from index.ts (TextStyles.css, Links.css)
- Cleaned up TextStyles.css by removing unused class definitions

Benefits:
- Consumers of H1-H6 and DecoratedLink components no longer pull in
  styled-components dependencies
- Better tree-shaking support for modern builds
- Clear separation between new (plain CSS) and legacy (styled-components) code
- Legacy exports remain available for backward compatibility with existing code
- Migration path is clearer: consumers can gradually migrate away from
  Typography.legacy.ts and Headings.legacy.ts


snaps

Update src/app/components/Typography/Links.css

Address remaining Copilot review comments: Fix export conflicts and side effects

Fixed 3 remaining unresolved review comments from Reviews #8 and #10:

**Problem**: Typography.legacy.ts imported linkColor/linkHover from Links.tsx,
which pulled in React and Links.css as module side effects. This undermined
the goal of keeping legacy styled-components helpers isolated.

**Solution**: Created Links.constants.ts as a side-effect-free constants module
that both Links.tsx and Typography.legacy.ts import from.

Benefits:
- No React imports in legacy module
- No CSS side effects in legacy module
- Clear separation of concerns: constants vs. components

**Problem**: index.ts used `export * from './Links'` and `export * from
'./Typography.legacy'`. Since both modules exported linkColor/linkHover, this
created duplicate export name collisions (TS2308) that could break TypeScript
builds.

**Solution**: Changed index.ts to selective export:
- Changed `export * from './Links'` to `export { DecoratedLink } from './Links'`
- linkColor/linkHover are now only exported from Typography.legacy.ts
- Avoids duplicate exports while maintaining backward compatibility

The comment in index.ts was already updated in a previous commit to clarify
that tree-shaking benefits require direct submodule imports (e.g.,
`./Headings`, `./Links`) rather than importing from the main entry point.

All changes maintain backward compatibility while improving module structure.


Address Copilot review comments: Fix comment clarity and link color duplication

1. Updated comment in index.ts to clarify that tree-shaking benefits require
   direct submodule imports (e.g., './Headings', './Links') rather than
   importing from the main entry point.

2. Fixed link color duplication by establishing Links.tsx as single source of
   truth for linkColor and linkHover constants:
   - Defined linkColor and linkHover constants in Links.tsx
   - Updated Links.css to use CSS variables (--link-color, --link-hover)
   - Bound CSS variables from JS constants in DecoratedLink component
   - Updated Typography.legacy.ts to re-export constants from Links.tsx

This eliminates the maintenance hazard of duplicated color values and ensures
consistency between CSS and JS constants.
