+++ src/app/components/Typography/Links.constants.ts
@@ -1,0 +1,13 @@
+/**
+ * Link Color Constants
+ *
+ * Single source of truth for link colors used across Typography components.
+ * These constants are:
+ * - Imported by Links.tsx and bound as CSS variables
+ * - Imported by Typography.legacy.ts for styled-components css fragments
+ *
+ * This module has no side effects (no React, no CSS imports).
+ */
+
+export const linkColor = '#027EB5';
+export const linkHover = '#0064A0';
