+++ src/app/components/Typography/Links.css
@@ -6,18 +6,18 @@
 
 /* Link style - standard link with underline */
 .link-style {
-  color: #027EB5;
+  color: var(--link-color, #027EB5);
   cursor: pointer;
   text-decoration: underline;
 }
 
 .link-style:hover {
-  color: #0064A0;
+  color: var(--link-hover, #0064A0);
 }
 
 /* Decorated link style - no underline by default, underline on hover/focus */
 .decorated-link-style {
-  color: #027EB5;
+  color: var(--link-color, #027EB5);
   cursor: pointer;
   text-decoration: none;
 }
@@ -25,7 +25,7 @@
 .decorated-link-style:hover,
 .decorated-link-style:focus {
   text-decoration: underline;
-  color: #0064A0;
+  color: var(--link-hover, #0064A0);
 }
 
 /* Decorated link disabled state */
@@ -35,5 +35,4 @@
   color: var(--text-color);
   cursor: not-allowed;
   opacity: 0.4;
-  text-decoration: none;
 }
